+++ src/Feature/Mfa/Bag/VerificationMfaBag.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smsapi\Client\Feature\Mfa\Bag;
+
+/**
+ * @api
+ */
+class VerificationMfaBag
+{
+    /**
+     * @var string
+     */
+    public $phoneNumber;
+    /**
+     * @var string
+     */
+    public $code;
+
+    public function __construct(string $code, string $phoneNumber)
+    {
+        $this->code = $code;
+        $this->phoneNumber = $phoneNumber;
+    }
+}