+++ src/Feature/Mfa/Data/Mfa.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smsapi\Client\Feature\Mfa\Data;
+
+/**
+ * @api
+ */
+class Mfa
+{
+    /**
+     * @var string
+     */
+    public $id;
+
+    /**
+     * @var string
+     */
+    public $code;
+
+    /**
+     * @var string
+     */
+    public $phoneNumber;
+
+    /**
+     * @var string
+     */
+    public $from;
+}