+++ src/Service/SmsapiComService.php
@@ -11,6 +11,7 @@ use Smsapi\Client\Feature\Ping\PingFeature;
 use Smsapi\Client\Feature\Profile\ProfileFeature;
 use Smsapi\Client\Feature\Push\PushFeature;
 use Smsapi\Client\Feature\ShortUrl\ShortUrlFeature;
+use Smsapi\Client\Feature\Mfa\MfaFeature;
 use Smsapi\Client\Feature\Sms\SmsFeature;
 use Smsapi\Client\Feature\Subusers\SubusersFeature;
 
@@ -22,6 +23,8 @@ interface SmsapiComService
     public function pingFeature(): PingFeature;
 
     public function smsFeature(): SmsFeature;
+
+    public function mfaFeature(): MfaFeature;
 
     public function hlrFeature(): HlrFeature;