+++ src/SmsapiClient.php
@@ -12,7 +12,7 @@ use Smsapi\Client\Service\SmsapiPlService;
  */
 interface SmsapiClient extends LoggerAwareInterface
 {
-    const VERSION = '2.5.0';
+    const VERSION = '2.6.0';
 
     public function setProxy(string $proxy): self;