+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2020-01-31
+### Added
+- generation and verification of MFA codes implement
+- `MfaFeature` uses mfa endpoint
+
 ## [2.5.0] - 2020-01-14
 ### Added
 - custom contact fields support